Fixed some compile warnings

+++ src/age/utils/mem_alloc.cpp
@@ -32,7 +32,7 @@
 extern "C" {
 #endif
 
-void * mmalloc(int size)
+void * mmalloc(size_t size)
 {
    void *p_tmp = malloc(size);
    if(!p_tmp) {
